modify controller error message and history MVC
