+++ controllers/eventController.js
@@ -1,6 +1,6 @@
 const eventModel = require("../models/eventModel");
 
-const getEvents = async (req, res) => {
+exports.getEvents = async (req, res) => {
   try {
     const page = parseInt(req.query.page) || 0;
     const PAGE_SIZE = 12;
@@ -8,7 +8,9 @@ const getEvents = async (req, res) => {
     const result = await eventModel.getEvents(page, PAGE_SIZE);
 
     if (result.length === 0) {
-      return res.status(404).json({ error: "No matching events found" });
+      return res
+        .status(404)
+        .json({ error: true, message: "No matching events found" });
     }
 
     // Check if there are more pages
@@ -22,13 +24,13 @@ const getEvents = async (req, res) => {
   }
 };
 
-const getEventById = async (req, res) => {
+exports.getEventById = async (req, res) => {
   try {
     const { id } = req.params; // Get the event ID from the URL parameters
     const result = await eventModel.getEventById(id);
 
     if (result.length === 0) {
-      return res.status(404).json({ error: "Event not found" });
+      return res.status(404).json({ error: true, message: "Event not found" });
     }
 
     const jsonResult = result[0]; // Only one event is expected
@@ -36,9 +38,4 @@ const getEventById = async (req, res) => {
   } catch (error) {
     res.status(500).json({ error: true, message: error.message });
   }
-};
-
-module.exports = {
-  getEvents,
-  getEventById,
 };
