+++ controllers/userController.js
@@ -1,7 +1,7 @@
 //userController
 const UserModel = require("../models/userModel");
 const bcrypt = require("bcryptjs");
-const { generateToken, verifyToken } = require("../utils/tokenUtils");
+const { generateToken } = require("../utils/tokenUtils");
 
 // Sign Up
 exports.signUp = async (req, res) => {
@@ -13,25 +13,29 @@ exports.signUp = async (req, res) => {
   // const password = req.body.password;
 
   if (!name || !email || !phone || !password) {
-    return res.status(400).json({ error: "資料欄位填寫有誤或留白" });
+    return res
+      .status(400)
+      .json({ error: true, message: "資料欄位填寫有誤或留白" });
   }
 
   try {
     const existingUser = await UserModel.findUserByEmail(email);
     if (existingUser.length > 0) {
-      return res.status(409).json({ error: "此email帳號已註冊" });
+      return res
+        .status(409)
+        .json({ error: true, message: "此email帳號已註冊" });
     }
 
     const existingPhone = await UserModel.findUserByPhone(phone);
     if (existingPhone.length > 0) {
-      return res.status(409).json({ error: "此手機號碼已註冊" });
+      return res.status(409).json({ error: true, message: "此手機號碼已註冊" });
     }
 
     await UserModel.createUser(name, email, phone, password);
     res.status(201).json({ message: "註冊成功" });
   } catch (error) {
     console.error("註冊時發生錯誤：", error);
-    res.status(500).json({ error: "伺服器發生錯誤" });
+    res.status(500).json({ error: true, message: "伺服器發生錯誤" });
   }
 };
 
@@ -40,14 +44,14 @@ exports.signIn = async (req, res) => {
   const { email, password } = req.body;
 
   if (!email || !password) {
-    return res.status(400).json({ error: "請填寫所有欄位" });
+    return res.status(400).json({ error: true, message: "請填寫所有欄位" });
   }
 
   try {
     const rows = await UserModel.findUserByEmail(email);
 
     if (rows.length === 0) {
-      return res.status(401).json({ error: "您輸入的帳號有誤" });
+      return res.status(401).json({ error: true, message: "您輸入的帳號有誤" });
     }
 
     const user = rows[0];
@@ -57,30 +61,29 @@ exports.signIn = async (req, res) => {
       const token = generateToken(user);
       res.status(200).json({ message: "登入成功", token });
     } else {
-      res.status(401).json({ error: "您輸入的密碼有誤" });
+      res.status(401).json({ error: true, message: "您輸入的密碼有誤" });
     }
   } catch (error) {
     console.error("登入時發生錯誤：", error);
-    res.status(500).json({ error: "伺服器發生錯誤" });
+    res.status(500).json({ error: true, message: "伺服器發生錯誤" });
   }
 };
 
 // New Get User Info
 exports.getUserInfo = async (req, res) => {
   try {
-    // 直接使用 req.user，這是 verifyToken 中間件解碼後存入的
     const email = req.user.email;
     const rows = await UserModel.getUserInfoByEmail(email);
 
     if (rows.length === 0) {
-      return res.status(404).json({ error: "查無使用者帳號" });
+      return res.status(404).json({ error: true, message: "查無使用者帳號" });
     }
 
     const user = rows[0];
     res.status(200).json(user);
   } catch (error) {
     console.error("錯誤", error);
-    res.status(500).json({ error: "伺服器發生錯誤" });
+    res.status(500).json({ error: true, message: "伺服器發生錯誤" });
   }
 };
 
@@ -109,7 +112,7 @@ exports.updatePhone = async (req, res) => {
     }
 
     await UserModel.updateUserPhone(userId, phone);
-    res.json({ success: true, message: "手機號碼更新成功" });
+    res.json({ message: "手機號碼更新成功" });
   } catch (error) {
     res.status(500).json({ error: true, message: error.message });
   }
@@ -121,7 +124,6 @@ exports.updatePassword = async (req, res) => {
     const userId = req.user.id;
     const { currentPassword, newPassword } = req.body;
 
-    // 從資料庫中獲取使用者資料
     const user = await UserModel.getUserById(userId);
     if (
       !user ||
@@ -132,9 +134,8 @@ exports.updatePassword = async (req, res) => {
         .json({ error: true, message: "Current password is incorrect" });
     }
 
-    // 更新密碼
     await UserModel.updateUserPassword(userId, newPassword);
-    res.json({ success: true, message: "Password updated successfully" });
+    res.json({ message: "Password updated successfully" });
   } catch (error) {
     console.error("Error updating password:", error);
     res.status(500).json({ error: true, message: error.message });
