+++ main.js
@@ -25,18 +25,16 @@ app.use(bodyParser.json()); // For parsing JSON bodies
 // Import routes
 const userRoutes = require("./routes/userRoutes");
 const eventRoutes = require("./routes/eventRoutes");
-// const areaRoutes = require("./routes/areaRoutes");
 const seatRoutes = require("./routes/seatRoutes");
 const orderRoutes = require("./routes/orderRoutes");
-const queryRoutes = require("./routes/queryRoutes");
+const historyRoutes = require("./routes/historyRoutes");
 
 // Use routes
 app.use("/api", userRoutes);
 app.use("/api", eventRoutes);
-// app.use("/api", areaRoutes);
 app.use("/api", seatRoutes);
 app.use("/api", orderRoutes);
-app.use("/api", queryRoutes);
+app.use("/api", historyRoutes);
 app.use("/node_modules", express.static(__dirname + "/node_modules"));
 
 app.get("/", function (req, res) {
